Remove RandomizerSkipException
This exception assumes that the randomizer implementation decides if the field
should be skipped or not. Or, in the randomizer interface, there is no context information to be able to make this decision. It is the role of the populator to decide about setting or skipping the field since it has all the context about the field being processed