move nextDouble method from AbstractRangeRandomizer to AbstractRandomizer
This change is necessary in order to be able to use the method by all
randomizers without creating a dependency between packages.

For example, the `StringRandomizer` currently has specific code to generate
the random String length in a integer range. It could be possible to
use a IntegerRangeRandomizer to re-use its logic, but this would create
a dependency between the `text` package and `range` package.

Moving `nextDouble` method to AbstractRandomizer not only removes code
duplication, but also makes it possible to inherit the method without
creating package dependencies.