issue-299: split argument value when argument type is array
Jackson is able to convert an array of String values to an
array of other Java types. However, it does not parse a raw
String and split it if it is intended to be mapped to an array.

This commit prepares the data for Jackson by splitting the
value of `RandomizerArgument` when its type is an array.

NB: The goal is to cover simple use cases of comma separated
values like reported in #299. There is no plan to make this
feature support custom separator, date format, number format, etc.