feat: introduce TypeResolver API
Introduces a new `TypeResolver` interface to the public API. This
interface's default implementation is in `ReflectionUtils`, which
is then delegated to the static helpers in `ClassGraphFacade`. This
means the new feature remains backwards compatible with previous
behavior.

Users can now specify a new parameter,
`EasyRandomParameters#typeResolver(TypeResolver)`, which lets them
specify how concrete types will be resolved without coupling the
library to `ClassGraph` in any way, though users can still use
`ClassGraph` to implement it themselves.

An internal proxy class is used to wrap custom instances of
`TypeResolver` set via the parameters. This is mostly because when
creating the default `ObjenesisObjectFactory`, we need to pass it the
current type resolver. However, it may get modified later on. So
instead, we pass a proxy, which we can always modify thereafter. The
proxy is not accessible to the user, so the only way to modify it is via
the parameters.